created account page
created account page with most of the functionality, there are some issues to fix. the buy_now view checks if auction.status != "active" and rejects payment. but won auctions that haven't been paid for are still marked as "active" and the auction is still marked as status="active" in the database even though the timer expired. Django doesn't automatically change the status when time runs out.