fixed the "pay now" in auctions won section
it used to show that the auction is not active, which is problematic considering it ended with a winner and the winner didn't pay for it, so it now shows the winner the payment page that uses the payment model, and actually pays for the order. however the pay now option remains after payment/order, so that issue needs to be fixed. it also appropriately shows it in the order history to confirm payment and shows "paid" in green.